Refactor command handler to seperate handling of commands into seperate modules
