+++ src/cli.rs
@@ -1,2 +1,4 @@
 pub mod cli_parser;
 pub mod command_handler;
+mod checkout;
+mod configure;
